Setup `~/.local/bin` to be in PATH, even in Bash
Assumed the entrypoint would still be `tat` as the first command
executed, which then activates the rest of the settings.

Still can't reliably change the user shell if the Ubuntu environment is
one where PAM is linked up to an LDAP server, since current shell is
dictated by what is written to a user's LDAP object. Same with password,
for that matter. This is the best we can consistently muster in the
meantime.